requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
Pillow==11.3.0
pdf2image==1.17.0
PyMuPDF==1.26.3
python-json-logger==3.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
aiofiles==24.1.0
pytest==8.4.1
pytest-asyncio==1.0.0
reportlab==4.1.0